refactor: extract voice input and wake-word into AppDelegate+Voice.swift
